feat: rebase and compress functions
