+++ components/sendCurrentPath.js
@@ -1,0 +1,5 @@
+import { stdout } from 'node:process';
+
+export default function sendCurrentPath (currentPath) {
+    stdout.write(`You are currently in ${currentPath}\n`);
+}
